fix(media): cast to_jsonb(:cnt::int) — asyncpg polymorphic type on retry backoff
_on_failure backoff-путь падал с DatatypeMismatchError: to_jsonb(:cnt)
не мог определить тип (asyncpg шлёт параметр как unknown). Этот путь
раньше не выполнялся (старый прод хот-лупил без retry), всплыло после
деплоя новой retry-логики. Каст :cnt::int чинит.

docs-not-needed

+++ vera3/services/media-worker/src/media_worker/__main__.py
@@ -235,7 +235,7 @@ async def _on_failure(event_id: int, meta: dict, err: str) -> str:
                 metadata = jsonb_set(
                   jsonb_set(
                     COALESCE(metadata, '{{}}'::jsonb),
-                    '{{media_retry_count}}', to_jsonb(:cnt)
+                    '{{media_retry_count}}', to_jsonb(:cnt::int)
                   ),
                   '{{media_next_retry_at}}',
                   to_jsonb((NOW() + INTERVAL '{backoff} minutes')::text)
